requirements.txt
python-dotenv
aiogram
langchain
langchain_groq
chromadb
datetime